fix: resolve server-only import conflicts with Pages Router and barrel exports
- Remove `import 'server-only'` from `endpoint-config.ts` as it gets included in client bundles via Pages Router (e.g., AMP pages), causing build failures.
- Extract `fetchYoutubeData` into a dedicated `utils/server/youtube.ts` file equipped with `server-only` to protect server-side fetching logic.
- Remove `fetchYoutubeData` from the shared `utils/index.ts` barrel export to prevent client components from inadvertently importing server-only modules.
- Add explanatory comments to `endpoint-config.ts` clarifying its runtime role and safe usage across environments.
